# Refactor `inlineOrLiftBinders`